requirements.txt
Django==4.2.1
Pillow==9.5.0
djangorestframework==3.14.0
django-recaptcha == 3.0.0